// src/constants/enum.ts
export enum TaskStatus {
  OPEN = "open",
  PENDING = "pending",
  CLAIM = "ready_to_claim",
  COMPLETED = "completed",
}
export enum SocialType {
  TelegramJoin = "telegram_join",
  YouTubeWatch = "youtube_watch",
  YoutubeLike = "youtube_like",
  YoutubeComment = "youtube_comment",
  InstagramFollow = "instagram_follow",
  InstagramComment = "instagram_comment",
  InstagramLike = "instagram_like",
  XShare = "x_share",
  XComment = "x_comment",
  XFollow = "x_follow",
  XLike = "x_like",
  TikTokLike = "tiktok_like",
  TikTokWatch = "tiktok_watch",
  TikTokComment = "tiktok_comment",
}
export enum ReferralType {
  Invite_5 = "invite_5",
  Invite_10 = "invite_10",
  Invite_20 = "invite_20",
}